requirements.txt
Flask==3.0.3
docx2pdf==0.1.8
python-docx==1.1.2
pdf2docx==0.5.8
Pillow==10.2.0
waitress==3.0.2